+++ game/game.cpp
@@ -233,9 +233,9 @@ public:
     hasCollided(false){}
     virtual ~TestObject(){}
 
-    void rectDraw(Area area, int portx, int porty, const Graphics::Bitmap & bmp, bool collision){
-        const int viewx = (area.x > portx ? area.x - portx : portx - area.x);
-        const int viewy = (area.y > porty ? area.y - porty : porty - area.y);
+    void rectDraw(Area area, double portx, double porty, const Graphics::Bitmap & bmp, bool collision){
+        const double viewx = (area.x > portx ? area.x - portx : portx - area.x);
+        const double viewy = (area.y > porty ? area.y - porty : porty - area.y);
         bmp.rectangle(viewx, viewy, viewx+area.width, viewy+area.height, 
                                                  (collision ? Graphics::makeColor(255, 0, 0) : Graphics::makeColor(128,128,128)));
     }
